requirements.txt
alembic==1.0.11
amqp==2.5.0
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.0.0
blinker==1.4
celery==4.3.0
Click==7.0
coverage==4.5.3
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.8.0
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.3
Mako==1.0.13
MarkupSafe==1.1.1
migrate==0.3.8
more-itertools==7.1.0
packaging==19.0
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.0
pytest==5.0.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
redis==3.2.1
six==1.12.0
SQLAlchemy==1.3.5
sqlalchemy-mptt==0.2.4
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.2
